requirements.txt
openai
python-dotenv
streamlit
audio-recorder-streamlit
whisper
